+++ packages/connectivity/src/index.ts
@@ -1,7 +1,7 @@
 import { makeEventListener } from "@solid-primitives/event-listener";
-import { createSharedRoot } from "@solid-primitives/rootless";
-import { createHydrateSignal } from "@solid-primitives/utils";
-import { Accessor, sharedConfig } from "solid-js";
+import { createHydratableSingletonRoot } from "@solid-primitives/rootless";
+import { createHydratableSignal, trueFn } from "@solid-primitives/utils";
+import { Accessor } from "solid-js";
 
 /**
  * Attaches event listeners and fires callback whenever `window.onLine` changes.
@@ -20,7 +20,10 @@ export function makeConnectivityListener(callback: (isOnline: boolean) => void):
   }
   const clear1 = makeEventListener(window, "online", callback.bind(void 0, true));
   const clear2 = makeEventListener(window, "offline", callback.bind(void 0, false));
-  return () => (clear1(), clear2());
+  return () => {
+    clear1();
+    clear2();
+  };
 }
 
 /**
@@ -33,25 +36,22 @@ export function makeConnectivityListener(callback: (isOnline: boolean) => void):
  */
 export function createConnectivitySignal(): Accessor<boolean> {
   if (process.env.SSR) {
-    return () => true;
+    return trueFn;
   }
-  const [status, setStatus] = createHydrateSignal<boolean>(true, () => navigator.onLine);
+  const [status, setStatus] = createHydratableSignal<boolean>(true, () => navigator.onLine);
   makeConnectivityListener(setStatus);
   return status;
 }
 
-const sharedConnectivitySignal = /*#__PURE__*/ createSharedRoot(createConnectivitySignal);
-
 /**
  * A signal representing the browser's interpretation of whether it is on- or offline.
  *
- * This is a [shared root primitive](https://github.com/solidjs-community/solid-primitives/tree/main/packages/rootless#createSharedRoot) except if during hydration.
+ * This is a [singleton root primitive](https://github.com/solidjs-community/solid-primitives/tree/main/packages/rootless#createSingletonRoot) except if during hydration.
  *
  * @return Returns a signal representing the online status. Read-only.
  * @example
  * const isOnline = useConnectivitySignal()
  * isOnline() // T: boolean
  */
-export const useConnectivitySignal: () => Accessor<boolean> = process.env.SSR
-  ? () => () => true
-  : () => (sharedConfig.context ? createConnectivitySignal() : sharedConnectivitySignal());
+export const useConnectivitySignal: () => Accessor<boolean> =
+  /*#__PURE__*/ createHydratableSingletonRoot(createConnectivitySignal);