+++ packages/fullscreen/src/index.ts
@@ -21,10 +21,10 @@ export const createFullscreen = (
   active?: Accessor<FullscreenOptions | boolean>,
   options?: FullscreenOptions,
 ): Accessor<boolean> => {
-  // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
   if (isServer) {
     return () => false;
   }
+
   const [isActive, setActive] = createSignal(false);
   createEffect(() => {
     if (ref) {