+++ site/src/components/Primitives/PrimitivePageMain.tsx
@@ -30,6 +30,7 @@ const PrimitivePageMain: ParentComponent<{
   return (
     <>
       <Title>{props.name}</Title>
+      <script>document.documentElement.classList.add("primitives-page-main")</script>
       <div
         class="-z-1 absolute top-0 left-0 right-0 h-[95vh] bg-[linear-gradient(to_bottom,#fff_var(--primitive-padding-top-gr),transparent)] dark:bg-[linear-gradient(to_bottom,#293843_var(--primitive-padding-top-gr),transparent)]"
         style={{